Unit1/FileIOAlgorithmTwo/src/us/andrearaujo/FileInput.java
package us.andrearaujo;

import java.io.BufferedReader;
import java.io.FileNotFoundException;
import java.io.FileReader;
import java.io.IOException;

/**
 * Created by mgreen14 on 12/27/17.
 * Taken over by andrearaujo on 9/5/18
 *
 * This class is for file input
 */
public class FileInput {

    private BufferedReader in = null;
    private String fileName;

    /**
     * The FileInput constructor requires the filename as a string
     *
     * @param fileName is the name of the file to read
     */
    public FileInput(String fileName) {
        this.fileName = fileName;
        try {
            in = new BufferedReader(new FileReader(fileName));
        } catch (FileNotFoundException e) {
            System.out.println("File Open Error: " + fileName + " " + e);
        }
    }

    /**
     * A void method that tries to read the file
     */
    public void fileRead() {
        String line;
        try {
            while ((line = in.readLine()) != null) {
                System.out.println(line);
            }
        } catch (Exception e) {
            System.out.println("File Write Error: " + fileName + " " + e);
        }
    }

    /**
     * A method to get a line from a file as a String
     *
     * @return line of text or null if there's an error
     */
    public String fileReadLine() {
        try {
            String line = in.readLine();
            return line;
        } catch (Exception e) {
            System.out.println("File Write Error: " + fileName + " " + e);
            return null;
        }
    }

    /**
     * Void method to close the file
     */
    public void fileClose() {
        if (in != null) {
            try {
                in.close();
            } catch (IOException e) {
                e.printStackTrace();
            }
        }

    }
}